# Conflicts:
#	app/config/routes.php
#	src/Controller/DBController.php
#	src/Model/Repository/UserModel.php
#	src/Model/Repository/UserTasks.php

+++ src/Controller/DBController.php
@@ -8,6 +8,7 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use SilexApp\Model\Repository\UserTasks;
 use Symfony\Component\HttpFoundation\RedirectResponse;
+use SilexApp\Model\Repository\EmailSender;
 
 
 class DBController
@@ -26,7 +27,7 @@ class DBController
             $response->setStatusCode(Response::HTTP_NOT_FOUND);
             $content = $app['twig']->render('LogIn.twig', [
                     'message' => 'User not found',
-                    'logejat' => false
+                    'logejat' => false,
                 ]
             );
             $response->setContent($content);
@@ -62,20 +63,25 @@ class DBController
         /** @var UploadedFile $img */
         $img = $request->files->get('newProfileImg');
 
-
         //$pass2 = htmlspecialchars($_POST['password2']);
         //$path = htmlspecialchars($_POST['files[]']);
 
         $repo = new UserTasks($app['db']);
-        $repo->deleteActualPic($name);
-        move_uploaded_file($img->getPathname(), './assets/uploads/' . $name . date("m-d-y"). date("h:i:sa") . ".jpg");
-        $img = './assets/uploads/' . $name . date("m-d-y"). date("h:i:sa") . ".jpg";
+        if ($img != NULL){
+            $repo->deleteActualPic($name);
+            move_uploaded_file($img->getPathname(), './assets/uploads/' . $name . date("m-d-y"). date("h:i:sa") . ".jpg");
+            $img = './assets/uploads/' . $name . date("m-d-y"). date("h:i:sa") . ".jpg";
+        }else{
+            $img = $repo->getActualProfilePic($name,$img);
+        }
 
         $repo->validateEditProfile($name, $birth, $pass1, $img);
         $response = new Response();
-        $content = $app['twig']->render('error.twig', [
+        $content = $app['twig']->render('editProfile.twig', [
                 'logejat' => true,
-                'message' => 'hola'
+                'username' => $name,
+                'birthdate' =>$birth,
+                'imagen' =>$img
             ]
         );
         $response->setContent($content);
@@ -91,145 +97,92 @@ class DBController
         $password = $request->get('password');
         /** @var UploadedFile $img */
         $img = $request->files->get('ProfileImg');
-        move_uploaded_file($img->getPathname(), './assets/uploads/' . $nickname . date("m-d-y"). date("h:i:sa") . ".jpg");
-        $img = './assets/uploads/' . $nickname . date("m-d-y"). date("h:i:sa") . ".jpg";
+        if ($img == NULL){
+            $img = './assets/img/' . "User.jpg";
+        }else{
+            move_uploaded_file($img->getPathname(), './assets/uploads/' . $nickname . date("m-d-y"). date("h:i:sa") . ".jpg");
+            $img = './assets/uploads/' . $nickname . date("m-d-y"). date("h:i:sa") . ".jpg";
+        }
+
         $repo = new UserTasks($app['db']);
         $exists = $repo->checkUser($nickname);
         $response = new Response();
 
 
         if (!$exists) {
-            $aleatorio = uniqid(); //Genera un id único para identificar la cuenta a traves del correo.
-
-            $mensaje = "Registro en tuweb.com\n\n";
-            $mensaje .= "Estos son tus datos de registro:\n";
-            $mensaje .= "Unsuario: $nickname.\n";
-            $mensaje .= "Contraseña: $password.\n\n";
-            $mensaje .= "Debes activar tu cuenta pulsando este enlace: grup5.dev/activacion.php?id=".$aleatorio;
-
-            $cabeceras = 'From: Dogygram@example.com' . "\r\n" .
-                'Reply-To: Dogygram@example.com' . "\r\n" .
-                'X-Mailer: PHP/' . phpversion();
-
-
-            //$sendMail = mail($email,'Activar cuenta',$mensaje, $cabeceras);
-            /*if($sendMail){*/
-                $repo->RegisterUser($nickname, $email, $birthdate, $password, $img);
-                $response->setStatusCode(Response::HTTP_OK);
-
-                $content = $app['twig']->render('error.twig', [
-                        'message' => 'Registro finalizado correctamente'. $img,
-                        'logejat' => false
-                    ]
-                );
-            /*}else{*/
-               // $response->setStatusCode(Response::HTTP_BAD_REQUEST);
-                //$content = $app['twig']->render('error.twig', [
-                  //      'message' => 'No se ha podido enviar el email',
-                    //    'logejat' => false
-
-                //]);
-            /*}
-            //$sendMail = mail($email,'Activar cuenta',$mensaje, $cabeceras);
-            if($sendMail){*/
+            //$sender = new EmailSender();
+            //if ($sender->sendEmail($email)){$repo->RegisterUser($nickname, $email, $birthdate, $password, $img);
             $repo->RegisterUser($nickname, $email, $birthdate, $password, $img);
             $response->setStatusCode(Response::HTTP_OK);
 
-            $sendMail = mail($email,'Activar cuenta',$mensaje, $cabeceras);
-            if($sendMail){
-                $repo->RegisterUser($nickname, $email, $birthdate, $password, $img);
-                $response->setStatusCode(Response::HTTP_OK);
-
-                $content = $app['twig']->render('validate.twig', [
-                    'message' => 'Activa tu usuario mediante el siguiente link:',
-                    'logejat' => false,
-                    'name' => $nickname
-                ]
-            );
-            }else{
-                $response->setStatusCode(Response::HTTP_BAD_REQUEST);
-                $content = $app['twig']->render('error.twig', [
-                        'message' => 'No se ha podido enviar el email',
-                        'logejat' => false
-
-                ]);
-            }
-        } else {
+            $content = $app['twig']->render('validate.twig', [
+                'message' => 'Email enviado correctamente:',
+                'logejat' => false,
+                'name' => $nickname,
+            ]);
+        }else{
             $response->setStatusCode(Response::HTTP_ALREADY_REPORTED);
             $content = $app['twig']->render('error.twig', [
-                    'message' => 'El usuario ya existe',
-                    'logejat' => false
+                     'message' => 'El usuario ya existe',
+                     'logejat' => false
                 ]
-            );
-        }
+              );
+            }
         $response->setContent($content);
 
         return $response;
-
-
     }
 
     public function DBnewPost(Application $app, Request $request)
     {
         $title = htmlspecialchars($request->get('title'));
-        //$imgName = htmlspecialchars($request->files->get('ProfileImg'));
         $privada = htmlspecialchars($request->get('privada'));
 
-        $size = $request->get('size');
-        //var_dump($path);
-
-
         /** @var UploadedFile $img */
-        $img = $request->files->get('images');
-        var_dump($img);
-        die;
-
-        move_uploaded_file($img->getPathname(), './assets/uploads/' . $title . date("m-d-y") .date("h:i:sa") . ".jpg");
-        $img = './assets/uploads/' . $title . date("m-d-y") .date("h:i:sa"). ".jpg";
-
-        //var_dump($privada);
-        //var_dump($request->files->get('imagen'));
-        //var_dump($request->files);
-        if ($privada === "on") {
-            $private = 1;
-        } else {
-            $private = 0;
-        }
-        if($size === 'gran'){
-            $sizeImage = 400;
-        }else{
-            $sizeImage = 100;
-        }
-
-        $folder = "./assets/img/";
-       // $path_name = $imgName;
-
-        //$folder = "/assets/img/";
-        //$path_name = $imgName;
-
-        //var_dump($path_name);
-
-
-        $repo = new UserTasks($app['db']);
-        $ok = $repo->DBnewPost($title, $img, $private, $sizeImage);
+        $img = $request->files->get('ProfileImg');
         $response = new Response();
+
         $repo = new UserTasks($app['db']);
-        if($app['session']->has('name')){
-            $log = true;
-        }
-
-        $usuari =  $app['session']->get('name');
-        $imgMesVistes = $repo->home1($log,$usuari);
-
-        if ($ok) {
-            $content = $app['twig']->render('hello.twig', [
-                    'logejat' => true,
-                    'dades' => $imgMesVistes
+        if ($img == NULL){
+            $usuari =  $app['session']->get('name');
+            $img = $repo ->getActualProfilePic($usuari, null);
+            $response->setStatusCode(Response::HTTP_NOT_FOUND);
+            $content = $app['twig']->render('newPost.twig', [
+                    'message' => 'IMAGE NOT FOUND',
+                    'logejat' => false,
+                    'username' => $usuari,
+                    'imagen' => $img
                 ]
             );
-        }
-        $response->setContent($content);
+            $response->setContent($content);
+            return $response;
+        }else{
+            move_uploaded_file($img->getPathname(), './assets/uploads/' . $title . date("m-d-y") .date("h:i:sa") . ".jpg");
+            $img = './assets/uploads/' . $title . date("m-d-y") .date("h:i:sa"). ".jpg";
 
+            if ($privada === "on") {
+                $private = 1;
+            } else {
+                $private = 0;
+            }
+
+            $ok = $repo->DBnewPost($title, $img, $private);
+            $repo = new UserTasks($app['db']);
+            if($app['session']->has('name')){
+                $log = true;
+            }
+            $usuari =  $app['session']->get('name');
+            $imgMesVistes = $repo->home1($log,$usuari);
+
+            if ($ok) {
+                $content = $app['twig']->render('hello.twig', [
+                        'logejat' => true,
+                        'dades' => $imgMesVistes
+                    ]
+                );
+            }
+            $response->setContent($content);
+        }
         return $response;
     }
 
@@ -253,7 +206,7 @@ class DBController
         } else {
             $response->setStatusCode(Response::HTTP_ALREADY_REPORTED);
             $content = $app['twig']->render('error.twig', [
-                    'message' => 'No se ha podido validar el usuario' . $nickname,
+                    'message' => 'No se ha podido validar el usuario ' . $nickname,
                     'logejat' => false
                 ]
             );
@@ -266,24 +219,26 @@ class DBController
     public function DBeditImage(Application $app, Request $request, $id)
     {
         $title = htmlspecialchars($request->get('title'));
-        $imgName = htmlspecialchars($request->files->get('imagen'));
+        $img = $request->files->get('imagen');
         $privada = htmlspecialchars($request->get('privada'));
-        $size = htmlspecialchars($request->get('size'));
-        if ($size === "gran"){
-            $sizeImage = 400;
-        }else{
-            $sizeImage = 100;
-        }
+
         if ($privada === "on") {
             $private = 1;
         } else {
             $private = 0;
         }
-        $folder = "/assets/img/";
-        $path_name = $imgName;
+
         //var_dump($path_name);
         $repo = new UserTasks($app['db']);
-        $repo->editInformation($title, $path_name, $private, $id, $sizeImage);
+
+        if ($img != NULL){
+            $repo->deleteActualPic($title);
+            move_uploaded_file($img->getPathname(), './assets/uploads/' . $title . date("m-d-y"). date("h:i:sa") . ".jpg");
+            $img = './assets/uploads/' . $title . date("m-d-y"). date("h:i:sa") . ".jpg";
+        }else{
+            $img = $repo->getActualPostImg($id,$img);
+        }
+        $repo->editInformation($title, $img, $private, $id);
         $dades = $repo->dadesImatges();
         $content = $app['twig']->render('galeria.twig', [
             'logejat' => true,
